+++ services/madoc-ts/src/routes/projects/get-project.ts
@@ -42,17 +42,19 @@ export const getProject: RouteMiddleware<{ id: string }> = async context => {
 
   const taskStatuses = taskStats.statuses || {};
 
-  statistics['0'] = content.canvases - taskStats.total;
-  statistics['1'] = taskStatuses['1'] || 0;
-  statistics['2'] = taskStatuses['2'] || 0;
-  statistics['3'] = taskStatuses['3'] || 0;
-
   let config;
   if (projectConfiguration.config && projectConfiguration.config[0] && projectConfiguration.config[0].config_object) {
     config = { ...staticConfiguration, ...projectConfiguration.config[0].config_object };
   } else {
     config = staticConfiguration;
   }
+
+  const countManifests = config.shadow?.showCaptureModelOnManifest;
+
+  statistics['0'] = countManifests ? content.manifests - taskStats.total : content.canvases - taskStats.total;
+  statistics['1'] = taskStatuses['1'] || 0;
+  statistics['2'] = taskStatuses['2'] || 0;
+  statistics['3'] = taskStatuses['3'] || 0;
 
   context.response.body = {
     ...project,