+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new zoom controls
 - Enabled new project template hook: `beforeCloneModel`
 - Enabled new project template hook: `onRevisionApproved`
+- Added support for Manifest capture models (only through project templates)
 
 ### Fixed
 - Regression with dropdown (reverts fix for overflow bug in popup)